pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-buttons"
version = "3.0.0"
description = "AI Share Buttons Generator - universal HTML snippets"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.28",
    "beautifulsoup4>=4.11",
    "lxml>=4.9",
    "pyyaml>=6.0",
]

[project.scripts]
ai-buttons = "ai_buttons.cli:main"

[tool.setuptools.packages.find]
include = ["ai_buttons*"]
